fix(review): cast fontSize to UInt32 for TypedAnswerState
Problem/需求: CI 再次报 cannot convert Int to UInt32 at ReviewSession:693,上次只修了 queuedClozeOrdinal 但漏了 fontSize 的类型不匹配。

Fix/方案: resolveTypedAnswerState 中 field.fontSize(Int) 传给 TypedAnswerState.fontSize(UInt32) 时显式 UInt32 转换。

复现路径(可选): GitHub Actions build log 报错 at AnkiApp/Sources/Review/ReviewSession.swift:693。

+++ AnkiApp/Sources/Review/ReviewSession.swift
@@ -657,7 +657,7 @@ final class ReviewSession {
                     placeholder: placeholder.rawToken,
                     expected: "",
                     fontName: "-apple-system",
-                    fontSize: 18,
+                    fontSize: 18 as UInt32,
                     combining: true
                 )
             }
@@ -690,7 +690,7 @@ final class ReviewSession {
                 placeholder: placeholder.rawToken,
                 expected: expected,
                 fontName: field.fontName,
-                fontSize: field.fontSize,
+                fontSize: UInt32(field.fontSize),
                 combining: placeholder.combining
             )
         } catch {
